addresses some review comments
